Refactor Startup.cs
