+++ Capri.Web/Startup.cs
@@ -11,6 +11,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.EntityFrameworkCore;
+using Capri.Web.Configuration;
 using Capri.Web.Services;
 using Capri.Web.Services.Settings;
 using Capri.Database;
@@ -20,6 +21,8 @@ namespace Capri.Web
 {
     public class Startup
     {
+        public IConfiguration Configuration { get; }
+
         public Startup(IConfiguration configuration)
         {
             Configuration = configuration;
@@ -30,45 +33,11 @@ namespace Capri.Web
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddMvc();
-            services.AddDbContext<CapriDbContext>(
-                options => options.UseSqlServer(Configuration["DbConnectionString"]));
-            services.AddScoped<ISqlDbContext, CapriDbContext>();
-
-            services.AddIdentity<User, GuidRole>(options => { })
-                    .AddEntityFrameworkStores<CapriDbContext>();
-
-            // configure strongly typed settings objects
-            var systemSettingsSection = Configuration.GetSection("SystemSettings");
-            var jwtSettingsSection = Configuration.GetSection("JwtSettings");
-            services.Configure<SystemSettings>(systemSettingsSection);
-            services.Configure<JwtSettings>(jwtSettingsSection);
-
-            // configure jwt authentication
-            var secret = jwtSettingsSection.Get<JwtSettings>().Secret;
-            var key = System.Text.Encoding.ASCII.GetBytes(secret);
-            services.AddAuthentication(x =>
-            {
-                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            })
-            .AddJwtBearer(x =>
-            {
-                x.RequireHttpsMetadata = false;
-                x.SaveToken = true;
-                x.TokenValidationParameters = new TokenValidationParameters
-                {
-                    ValidateIssuerSigningKey = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(key),
-                    ValidateIssuer = false,
-                    ValidateAudience = false
-                };
-            });
-
-            // configure DI for application services
+            services.AddDatabaseConfiguration();
+            services.AddJwtConfiguration();
+            services.AddSystemConfiguration();
             services.AddScoped<ILoginService, LoginService>();
         }
-
-        public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
